+++ pkg/bbui-0.9.5.js
@@ -811,7 +811,8 @@ bb.actionBar = {
 		if (actionBar.hasAttribute('data-bb-back-caption') && actionBar.querySelectorAll('[data-bb-style=tab]').length == 0) {		
 			var chevron,
 				backCaption,
-				backslash;
+				backslash,
+				backHighlight;
 			backBtn = document.createElement('div');
 			backBtn.setAttribute('class','bb-bb10-action-bar-back-button-'+res+' bb-bb10-action-bar-back-button-'+res+'-' + color);
 			backBtn.onclick = bb.popScreen;
@@ -825,6 +826,23 @@ bb.actionBar = {
 			backCaption.setAttribute('class','bb-bb10-action-bar-back-text-'+res);
 			backCaption.innerHTML = actionBar.getAttribute('data-bb-back-caption');
 			backBtn.appendChild(backCaption);
+			// Create our highlight for touch
+			backHighlight = document.createElement('div');
+			backHighlight.setAttribute('class','bb-bb10-action-bar-back-button-highlight');
+			backHighlight.style['position'] = 'absolute';
+			backHighlight.style['height'] = bb.device.isPlayBook ? '57px' : '110px';
+			backHighlight.style['width'] = bb.device.isPlayBook ? '4px' : '8px';
+			backHighlight.style['background-color'] = 'transparent';
+			backHighlight.style['top'] = bb.device.isPlayBook ? '8px' : '15px';
+			backBtn.backHighlight = backHighlight;
+			backBtn.appendChild(backHighlight);
+			backBtn.ontouchstart = function() {
+					this.backHighlight.style['background-color'] = bb.options.highlightColor;				
+			}
+			backBtn.ontouchend = function() {
+					this.backHighlight.style['background-color'] = 'transparent';				
+			}
+			
 			// Create our backslash
 			backslash = document.createElement('div');
 			backslash.setAttribute('class','bb-bb10-action-bar-back-slash-'+res+'-'+color); 
@@ -955,6 +973,10 @@ bb.actionBar = {
 									// Compute margins
 									margins = (actionType == 'tab') ? 2 : 0;
 									action.style.width = (actionWidth - margins) + 'px'; 
+									if (action.highlight) {
+										action.highlight.style['width'] = (actionWidth * 0.6) + 'px';
+										action.highlight.style['margin-left'] = (actionWidth * 0.2) + 'px';
+									}
 								}
 								// Adjust our more button
 								if (this.moreBtn && (this.shownActions.length > 0)) {
@@ -1169,6 +1191,15 @@ bb.actionBar = {
 					action.normal = 'bb-bb10-action-bar-action-'+res+' bb-bb10-action-bar-button-'+color;
 				}
 				
+				// Highlight on touch
+				action.ontouchstart = function() {
+						this.highlight.style['background-color'] = bb.options.highlightColor;				
+				}
+				// Remove highlight when touch ends
+				action.ontouchend = function() {
+						this.highlight.style['background-color'] = 'transparent';				
+				}
+				
 				// Assign the setCaption function
 				action.setCaption = function(value) {
 									this.display.innerHTML = value;
@@ -1181,11 +1212,9 @@ bb.actionBar = {
 								};
 				action.setImage = action.setImage.bind(action);
 			}
-			
-			
+			// Default settings
 			action.innerHTML = '';
 			action.setAttribute('class',action.normal);
-			
 			action.appendChild(icon);
 			lastStyle = 'button';
 			
@@ -1194,7 +1223,18 @@ bb.actionBar = {
 			display.setAttribute('class','bb-bb10-action-bar-action-display-'+res);
 			display.innerHTML = caption;
 			action.display = display;
-			action.appendChild(display);	
+			action.appendChild(display);
+
+			// Set our highlight
+			if (action.getAttribute('data-bb-img') != 'overflow') {
+				action.highlight = document.createElement('div');
+				action.highlight.setAttribute('class','bb-bb10-action-bar-action-highlight');
+				action.highlight.style['height'] = bb.device.isPlayBook ? '4px' : '8px';
+				action.highlight.style['width'] = (btnWidth * 0.6) + 'px';
+				action.highlight.style['margin-left'] = (btnWidth * 0.2) + 'px';
+				action.highlight.style['background-color'] = 'transparent';
+				action.appendChild(action.highlight);
+			}
 		}
 		// Center the action overflow items
 		if (actionBar.menu) {
@@ -2120,7 +2160,7 @@ bb.screen = {
 				
 				// Inner Scroll Area
 				scrollArea = document.createElement('div');
-				outerScrollArea.appendChild(scrollArea); 			
+				outerScrollArea.appendChild(scrollArea); 				
 				
 				// Copy all nodes in the screen that are not the action bar
 				for (j = 0; j < outerElement.childNodes.length - 1; j++) {
@@ -3964,6 +4004,7 @@ _bb10_grid = {
 					else if (type == 'row') {
 						var k,
 							table,
+							columnCount = 0,
 							tr,
 							td,
 							numItems,
@@ -3975,10 +4016,16 @@ _bb10_grid = {
 							height,
 							width,
 							hasOverlay,
+							hardCodedColumnNum = -1, // none specified
 							rowItems = innerChildNode.querySelectorAll('[data-bb-type=item]');
 						
 						numItems = rowItems.length;
 						if (numItems == 0) continue;
+						
+						// See if they specified the number of items per column
+						if (innerChildNode.hasAttribute('data-bb-columns')) {
+							hardCodedColumnNum = innerChildNode.getAttribute('data-bb-columns');
+						}
 						
 						table = document.createElement('table');
 						table.style.width = '100%';
@@ -3986,8 +4033,22 @@ _bb10_grid = {
 						tr = document.createElement('tr');
 						table.appendChild(tr);
 
+						// Calculate the width
+						if (hardCodedColumnNum > 0) {
+							width = (window.innerWidth/hardCodedColumnNum) - 6;
+						} else {
+							width = (window.innerWidth/numItems) - 6;
+						}
+							
 						for (k = 0; k < numItems; k++) {
 							itemNode = rowItems[k];
+							
+							// Do not show more than the hardcoded number of items
+							if ((hardCodedColumnNum > 0) && ((k-1) > hardCodedColumnNum)) {
+								itemNode.style.display = 'none';
+								continue;
+							}
+							
 							subtitle = itemNode.innerHTML;
 							title = itemNode.getAttribute('data-bb-title');
 							hasOverlay = (subtitle || title);
@@ -3996,8 +4057,8 @@ _bb10_grid = {
 							td = document.createElement('td');
 							tr.appendChild(td);
 							td.appendChild(itemNode);
-							// deal with our margins
-							width = (window.innerWidth/numItems) - 6;
+							columnCount++;
+							
 							// Find out how to size the images
 							if (outerElement.isSquare) {
 								height = width;
@@ -4053,7 +4114,8 @@ _bb10_grid = {
 													};
 							itemNode.ontouchend = function() {
 														if (this.overlay) {
-															this.overlay.style = '';
+															this.overlay.style['opacity'] = '';
+							                                this.overlay.style['background-color'] = '';
 														}
 														itemNode.fingerDown = false;
 														if (itemNode.contextShown) {
@@ -4069,6 +4131,18 @@ _bb10_grid = {
 														};
 							itemNode.touchTimer = itemNode.touchTimer.bind(itemNode);
 						}
+						
+						// if there were hardcoded columns and not enough items to fit those columns, add the extra columns
+						if ((hardCodedColumnNum > 0) && (columnCount < hardCodedColumnNum)) {
+							var diff = hardCodedColumnNum - columnCount;
+							innerChildNode.extraColumns = [];
+							for (k = 0; k < diff; k++) {
+								td = document.createElement('td');
+								tr.appendChild(td);
+								td.style.width = width + 'px';
+								innerChildNode.extraColumns.push(td);
+							}
+						}
 					}
 				}
 			}
@@ -4078,17 +4152,33 @@ _bb10_grid = {
 									var items = this.querySelectorAll('[data-bb-type=row]'),
 										i,j,
 										rowItems,
+										row,
 										numItems,
 										itemNode,
 										width,
 										height;
 				
 									for (i = 0; i < items.length; i++) {
-										rowItems = items[i].querySelectorAll('[data-bb-type=item]');
+										var hardCodedColumnNum = -1;
+										row = items[i];
+										rowItems = row.querySelectorAll('[data-bb-type=item]');
 										numItems = rowItems.length;
+										
+										// See if they specified the number of items per column
+										if (row.hasAttribute('data-bb-columns')) {
+											hardCodedColumnNum = row.getAttribute('data-bb-columns');
+										}
+										
+										// Calculate the width
+										if (hardCodedColumnNum > 0) {
+											width = (window.innerWidth/hardCodedColumnNum) - 6;
+										} else {
+											width = (window.innerWidth/numItems) - 6;
+										}										
+										
+										// Adjust all the items
 										for (j = 0; j < numItems; j++ ) {
 											itemNode = rowItems[j];
-											width = (window.innerWidth/numItems) - 6;
 											if (outerElement.isSquare) {
 												height = width;
 											} else {
@@ -4107,6 +4197,13 @@ _bb10_grid = {
 											itemNode.style['-webkit-transition-duration'] = '0.2s';
 											itemNode.style['-webkit-transition-timing-function'] = 'linear';
 											itemNode.style['-webkit-transform'] = 'translate3d(0,0,0)';
+										}
+										
+										// Adjust the extra columns if there was hard coded columns that were not filled
+										if (row.extraColumns) {
+											for (j = 0; j < row.extraColumns.length;j++) {
+												row.extraColumns[j].style.width = width + 'px';
+											}
 										}
 									}
 								};