+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Below you will find all the different changes that have been added since the fir
 
 ## Version 0.9.5
 
+* BlackBerry 10 Styling
+	* Updated Action Bar back button and "button actions" to have a highlight when touched
+	* Updated grids to allow for a specified number of columns in a row
 * JavaScript Interfaces
 	* **Checkboxes:** Added show(), hide(), remove(), enable(), disable() functions 
 	* **Toggle Buttons:** Added show(), hide(), remove() functions